backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pydantic-settings
python-multipart
